% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_most_recent_date}
\alias{get_most_recent_date}
\title{Get date for most recent day of the week}
\usage{
get_most_recent_date(day)
}
\arguments{
\item{day}{character string in format of "Monday", "Tuesday", etc.}
}
\value{
date object
}
\description{
This function gets the most recent date for the day of the week.
This can be useful in these types of contexts...

* `map()` returns a list or a data frame

* `map_lgl()`, `map_int()`, `map_dbl()` and `map_chr()` return
    vectors of the corresponding type (or die trying);

* `map_dfr()` and `map_dfc()` return data frames created by row-binding
    and column-binding respectively. They require dplyr to be installed.
}
\examples{
get_most_recent_date("Monday")
}
